+++ libraryBackend/libraryProject/settings.py
@@ -42,8 +42,8 @@ INSTALLED_APPS = [
     "rest_framework.authtoken",
     "drf_yasg",
     "knox",
-    # "libraryApp", #.apps.LibraryappConfig",
-    "Account",
+    "libraryApp", #.apps.LibraryappConfig",
+    #"Account",
 ]
 
 MIDDLEWARE = [
@@ -141,7 +141,7 @@ USE_TZ = True
 # https://docs.djangoproject.com/en/3.1/howto/static-files/
 
 STATIC_URL = "/static/"
-AUTH_USER_MODEL = "Account.User"
+AUTH_USER_MODEL = "libraryApp.User"
 
 REST_FRAMEWORK = {
     "DEFAULT_AUTHENTICATION_CLASSES": [
